+++ circuits/environment/src/traits/eject.rs
@@ -88,7 +88,7 @@ fn eject_mode(start_mode: Mode, modes: &[Mode]) -> Mode {
 /****** Arrays ******/
 /********************/
 
-impl<C: Eject<Primitive = P>, P: Default> Eject for Vec<C> {
+impl<C: Eject<Primitive = P>, P> Eject for Vec<C> {
     type Primitive = Vec<P>;
 
     /// A helper method to deduce the mode from a list of `Eject` circuits.
@@ -104,7 +104,7 @@ impl<C: Eject<Primitive = P>, P: Default> Eject for Vec<C> {
     }
 }
 
-impl<C: Eject<Primitive = P>, P: Default, const N: usize> Eject for [C; N] {
+impl<C: Eject<Primitive = P>, P, const N: usize> Eject for [C; N] {
     type Primitive = Vec<P>;
 
     /// A helper method to deduce the mode from a list of `Eject` circuits.
@@ -120,7 +120,7 @@ impl<C: Eject<Primitive = P>, P: Default, const N: usize> Eject for [C; N] {
     }
 }
 
-impl<C: Eject<Primitive = P>, P: Default> Eject for &[C] {
+impl<C: Eject<Primitive = P>, P> Eject for &[C] {
     type Primitive = Vec<P>;
 
     /// A helper method to deduce the mode from a list of `Eject` circuits.
